Add option to pass debug info to Winscope through cross-tool protocol
This will be used by ATI to pass debug information that can we useful in Winscope to automate some debugging steps and highlight issues when opening Winscope traces from an ATI test failure.

For now we will just process Flicker failure stacktraces to jump to the
timestamp of the Flicker failure.

Test: winscope presubmit + local manual testing with ATI
Bug: 397414940
Change-Id: Id931a74ed93d36a5521b9ee9aeadba0334a11598