Make the crates.io fetcher work with rust 1.85
Rust 1.85 uses a different hasher, so the crates.io index
directory, ~/.cargo/registry/index/index.crates.io-<HASH> is different.

The crates_index crate does not automatically take the rust version into
account when computing the directory name, and defaults to using the
pre-1.85 hasher.

Also, when suggesting updates, keep going when we get an error for a crate
when trying to fetch data from crates.io (which was how the directory
name problem manifests, and was discovered).

Test: RUST_LOG=debug ./crate_tool suggest-updates --json --patches --semver-compatibility=ignore
Change-Id: I7d9af9270141c489b6bfaa4cc42b9545383155d8