Improve AudioInjector restart latency
When the VDM AudioInjector receives audio frames to inject, the audio
buffer of the remote submix associated AudioTrack is flushed of previous
written silence. In order for this to be effective, the state of the
track needs to be != ACTIVE, so pause the track, flush it and start it
again.

Bug: 435210862
Test: local VdmDemo sample app
Flag: EXEMPT, only demo sample app
Change-Id: I1a7d916f0cc6aaafbb7b7d03c83706073204ccea