Styling changes for pinned nodes/rects.
No need to scale pinned borders by the transform scale, this leads to
very thick borders for no reason in the 3D view.

Also removed the padding and leaf icon on the pinned nodes to allow the
background color to fill the width of the node on hover/select, these
nodes don't need to be aligned with the main tree as they are
specifically pinned.

Bug: 442786782
Test: manual
Change-Id: I282efe3b51e38092d6d70fb1239f60e00771e25c